refactoring: Move data-related stuff to /data
