+++ devito/__init__.py
@@ -8,7 +8,7 @@ from itertools import product
 import cpuinfo
 
 from devito.base import *  # noqa
-from devito.data import *  # noqa
+from devito.data.allocators import *  # noqa
 from devito.dimension import *  # noqa
 from devito.equation import *  # noqa
 from devito.grid import *  # noqa
