+++ devito/data/__init__.py
@@ -1,0 +1,3 @@
+from devito.data.allocators import *  # noqa
+from devito.data.decomposition import *  # noqa
+from devito.data.data import *  # noqa
